Ball changes colour based on proximity to one of four teams
The radio broadcasters can now be switched between broadcasting one of
four team "names" ("c", "m", "y", and "p").

The ball constantly records the strengths of the radio signals it
receives from the team broadcasters, and every 200 milliseconds it
displays the "winning" team on its display, and updates the neopixels
to the appropriate team colour.

This seems to work fairly well, but the neopixels can get a bit
"twitchy" (rapidly switching between colours) when the ball is
equally far away from two or more broadcasters, or when a person is
blocking the signal from the nearest broadcaster.